RVA & SPARQL all pushed back to Source
